separacion por estructuras clientes -cuentas - movimientos, y por funciones de utilidades e interfaz
