+++ interfaz.h
@@ -1,0 +1,32 @@
+#ifndef INTERFAZ_H_INCLUDED
+#define INTERFAZ_H_INCLUDED
+
+
+#include <windows.h>
+#include <stdio.h>
+#include <string.h>
+#include "gotoxy.h"  // usa gotoxy(), color(), etc.
+
+// === COLORES NOMBRADOS ===
+void colorNormal();
+void colorTitulo();
+void colorMenu();
+void colorError();
+void colorDato();
+void colorLinea();
+
+// === SONIDOS ===
+void beepOk();
+void beepError();
+void beepMenu();
+void beepConfirm();
+
+// === FUNCIONES VISUALES ===
+void limpiarPantalla();
+void printCenter(int y, const char* texto);
+void printLinea(int y);
+void printMarco();
+void esperaTecla();
+
+
+#endif // INTERFAZ_H_INCLUDED
