+++ main.c.save
@@ -1,10 +1,18 @@
-#include "clientes.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
 #include <time.h>
 #include "gotoxy.h"
 #include <windows.h> // Para Sleep
+#include "interfaz.h"
+#include "utils.h"
+
+#include "clientes.h"
+#include "cuentas.h"
+
+
+
+
 
 // Prototipos de los submenús
 void abmClientes();
@@ -102,7 +110,7 @@ void abmMovimientos() {
         menuMovimientos();
         opcion = getch();
         beepMenu();
-        switch(opcion) { 
+        switch(opcion) {
             case '1': altaMovimiento(); break;
             case '2': bajaMovimiento(); break;
             case '3': modificarMovimiento(); break;
